fix response time calculation, fixed and added scripts for graph

+++ PA2/FCFS_implementation.c
@@ -52,8 +52,8 @@ struct Process {
     pid_t pid;
     int *running;
     const char *name;
-    struct timeval creation_time;
-    struct timeval first_run_time;
+    struct timeval start_time;
+    struct timeval completion_time;
 };
 
 long get_elapsed_time(struct timeval start, struct timeval end) {
@@ -96,26 +96,31 @@ int main(int argc, char const *argv[])
     kill(pid4, SIGSTOP);
 
     struct Process processes[] = {
-        {pid1, &running1, "WORKLOAD1", current_time, {0}},
-        {pid2, &running2, "WORKLOAD2", current_time, {0}},
-        {pid3, &running3, "WORKLOAD3", current_time, {0}},
-        {pid4, &running4, "WORKLOAD4", current_time, {0}}
+        {pid1, &running1, "WORKLOAD1", {0}, {0}},
+        {pid2, &running2, "WORKLOAD2", {0}, {0}},
+        {pid3, &running3, "WORKLOAD3", {0}, {0}},
+        {pid4, &running4, "WORKLOAD4", {0}, {0}}
     };
     int num_processes = sizeof(processes) / sizeof(processes[0]);
 
     for (int i = 0; i < num_processes; i++) {
-        gettimeofday(&processes[i].first_run_time, NULL);
-        long response_time = get_elapsed_time(processes[i].creation_time, processes[i].first_run_time);
-        printf("Response time for %s: %ld microseconds\n", processes[i].name, response_time);
-
         printf("Starting process %s\n", processes[i].name);
         
+        gettimeofday(&processes[i].start_time, NULL);  // Record start time
+        kill(processes[i].pid, SIGCONT);
+
         while (*processes[i].running > 0) {
-            kill(processes[i].pid, SIGCONT);
             usleep(1000);
             kill(processes[i].pid, SIGSTOP);
             waitpid(processes[i].pid, processes[i].running, WNOHANG);
+            if (*processes[i].running > 0) {
+                kill(processes[i].pid, SIGCONT);
+            }
         }
+        
+        gettimeofday(&processes[i].completion_time, NULL);  // Record completion time
+        long response_time = get_elapsed_time(processes[i].start_time, processes[i].completion_time);
+        printf("Response time for %s: %ld microseconds\n", processes[i].name, response_time);
         
         printf("Process %s has completed\n", processes[i].name);
     }
